Finished Stand Back Up! How to Reverse Ragdoll Physics
Finished
Stand Back Up! How to Reverse Ragdoll Physics (Unity Tutorial)
by
Ketra Games

End of video mentions fixing the issue when the zombie lands on their stomach during ragdoll, and the jarring transition between that and being on their back.
Says it will be addressed in a future video.